pyproject.toml
[tool.black]
# Black 코드 포맷터 설정
line-length = 88
target-version = ['py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
# isort import 정렬 설정
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["app", "core", "services", "schemas", "middleware"]
sections = ["FUTURE", "STDLIB", "THIRDPARTY", "FIRSTPARTY", "LOCALFOLDER"]
combine_as_imports = true
force_grid_wrap = 0
include_trailing_comma = true
use_parentheses = true

[tool.mypy]
# MyPy 타입 체크 설정
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
strict_optional = true
no_implicit_reexport = true
show_error_codes = true

# 외부 라이브러리에 대한 타입 체크 무시
[[tool.mypy.overrides]]
module = [
    "google.generativeai.*",
    "mangum.*",
    "slowapi.*"
]
ignore_missing_imports = true

[tool.pytest.ini_options]
# Pytest 테스트 설정
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=.",
    "--cov-report=term-missing",
    "--cov-report=html:htmlcov",
    "--cov-fail-under=80"
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests"
]
asyncio_mode = "auto"

[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-political-columnist"
version = "1.0.0"
description = "AI 정치 컬럼니스트 - FastAPI 기반 정치 컬럼 생성 서비스"
authors = [
    {name = "AI Political Columnist Team", email = "team@example.com"}
]
license = {text = "MIT"}
readme = "README_PYTHON.md"
requires-python = ">=3.11"
keywords = ["fastapi", "ai", "gemini", "political", "columnist"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Framework :: FastAPI",
]

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "mangum>=0.19.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "google-generativeai>=0.8.0",
    "slowapi>=0.1.9",
    "python-multipart>=0.0.17",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "black>=24.10.0",
    "isort>=5.13.0",
    "flake8>=7.1.0",
    "mypy>=1.14.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.25.0",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.0",
]

[project.urls]
Homepage = "https://github.com/example/ai-political-columnist"
Repository = "https://github.com/example/ai-political-columnist.git"
Issues = "https://github.com/example/ai-political-columnist/issues"